+++ src/lib/interface/iContentBlock.ts
@@ -1,7 +1,0 @@
-export interface iContentBlock {
-    custom_id: string;
-    category: string;
-    content: string;
-    label?: string;
-    tokens: string[];
-}