+++ src/lib/store/contentBlocks.ts
@@ -1,4 +1,0 @@
-import type { iContentBlock } from '$lib/interface/iContentBlock.js';
-import { writable } from 'svelte/store';
-
-export const contentBlocks = writable([] as iContentBlock[]);