+++ src/lib/store/currentlyLoadedLocale.ts
@@ -1,8 +1,0 @@
-import { writable, type Writable } from 'svelte/store';
-
-/**
- * This only changes when a new set of translations are fully loaded, processed and ready for use.
- */
-export const currentlyLoadedLocale: Writable<string> = writable('');
-
-export default currentlyLoadedLocale;